Separate dolos class into dataset class in library
